Reset Ventialtion.Schedule.id properly so set WUFI RoomVentilation
